Revert "freebsd,linux: add recvmmsg() + sendmmsg() udp implementation"
This reverts commit 3d7136639a39db46bc4a9074922559a564e49514.
This reverts commit d9cd7d437d6bcea56355b6e0ef215aa64687d7a1.

The first reverted commit is the sendmmsg/recvmmsg support, the second
one a fix-up to deliver datagrams in order. The second commit has been
implicated in causing bind9 to crash on freebsd.

A quick review of the code suggests that downstream code written against
pre-v1.35.0 libuv can't safely deal with multi-datagram support because
they are unaware of the `UV_UDP_MMSG_CHUNK` flag and what that implies
for buffer management, hence I'm moving to revert it.

The `UV_UDP_MMSG_CHUNK` flag remains part of `uv_udp_flags` for
API/ABI backwards compatibility reasons but it is no longer used.

Fixes: https://github.com/libuv/libuv/issues/2791
Refs: https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=245653
